Update README.md
Assignment completion for a Responsive Resume. 
Created my Resume using the HTML and CSS properties.
Created my Resume to reflect my skills and understand the basic concepts of the HTML and CSS.
While creating the same learn to create the classes properly to give the stylings to them.
Also learned different CSS selectors and used many of them in my Responsive Resume.
As of now just created the for the practice purpose and later on will adding new things to the same resume.